rename FeedMewWrapper to FeedMew
